Move the SortCodeInput component to its own file
